requirements.txt
django==1.8.4
djangorestframework==3.1.3
django-websocket-redis==0.4.4
django-redis-cache==1.6.0
django-redis-sessions=0.5.0